app/src/requirements.txt
Click==8.1.3
gunicorn==21.2.0
itsdangerous==2.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.0

jinja2==3.1.2
flask==3.0.0
pymongo==3.12.3
flask_restful==0.3.9
requests==2.28.0
flask_cors==3.0.10
flask_socketio==5.3.6
pyarrow==14.0.2
tabulate==0.8.9
boto3==1.21.32
greenlet==3.0.3
gevent==23.9.1
flask-executor==1.0.0
newrelic==8.8.1
protobuf==4.25.2
scipy==1.7.3
pandas==1.4.2
scikit-learn==1.0.2
LightGBM==4.2.0
geopy==2.4.1
redispy==3.0.0
